wip: add serialization support
