+++ nodes/standalone/src/rpc.rs
@@ -28,6 +28,8 @@ use std::sync::Arc;
 use jsonrpsee::RpcModule;
 
 use pallet_did_lookup::linkable_account::LinkableAccountId;
+use public_credentials::CredentialEntryOf;
+use mashnet_node_runtime::Runtime;
 use runtime_common::{AccountId, Balance, Block, BlockNumber, DidIdentifier, Hash, Index};
 pub use sc_rpc_api::DenyUnsafe;
 use sc_transaction_pool_api::TransactionPool;
@@ -55,9 +57,11 @@ where
 	C::Api: pallet_transaction_payment_rpc::TransactionPaymentRuntimeApi<Block, Balance>,
 	C::Api: BlockBuilder<Block>,
 	C::Api: did_rpc::DidRuntimeApi<Block, DidIdentifier, AccountId, LinkableAccountId, Balance, Hash, BlockNumber>,
+	C::Api: public_credentials_rpc::PublicCredentialsRuntimeApi<Block, String, Hash, CredentialEntryOf<Runtime>>,
 	P: TransactionPool + 'static,
 {
 	use did_rpc::{DidApiServer, DidQuery};
+	use public_credentials_rpc::{PublicCredentialsApiServer, PublicCredentialsQuery};
 	use frame_rpc_system::{System, SystemApiServer};
 	use pallet_transaction_payment_rpc::{TransactionPayment, TransactionPaymentApiServer};
 
@@ -77,6 +81,7 @@ where
 	//
 	// `module.merge(YourRpcStruct::new(ReferenceToClient).into_rpc())?;`
 	module.merge(DidQuery::new(client).into_rpc())?;
+	module.merge(PublicCredentialsQuery::new(client).into_rpc())?;
 
 	Ok(module)
 }
