+++ pallets/public-credentials/Cargo.toml
@@ -60,6 +60,7 @@ std = [
   "log/std",
   "pallet-balances/std",
   "scale-info/std",
+  "serde",
   "sp-core/std",
   "sp-io/std",
   "sp-runtime/std",
