+++ pallets/public-credentials/src/credentials.rs
@@ -76,6 +76,8 @@ pub struct Credential<
 /// block. The block number is used to query the full content of the credential
 /// from archive nodes.
 #[derive(Encode, Decode, Clone, MaxEncodedLen, RuntimeDebug, PartialEq, Eq, PartialOrd, Ord, TypeInfo)]
+#[cfg_attr(feature = "std", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "std", serde(bound = ""))]
 #[scale_info(skip_type_params(T))]
 #[codec(mel_bound())]
 pub struct CredentialEntryOf<T: Config> {
