+++ support/src/deposit.rs
@@ -26,20 +26,8 @@ use sp_runtime::{traits::Zero, DispatchError};
 /// An amount of balance reserved by the specified address.
 #[derive(Clone, Debug, Encode, Decode, Eq, PartialEq, Ord, PartialOrd, TypeInfo, MaxEncodedLen)]
 #[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-#[cfg_attr(
-	feature = "std",
-	serde(bound(
-		serialize = "
-		Balance: std::fmt::Display,
-		Account: Serialize",
-		deserialize = "
-		Balance: std::str::FromStr,
-		Account: Deserialize<'de>"
-	))
-)]
 pub struct Deposit<Account, Balance> {
 	pub owner: Account,
-	#[cfg_attr(feature = "std", serde(with = "serde_balance"))]
 	pub amount: Balance,
 }
 
@@ -57,23 +45,6 @@ pub fn reserve_deposit<Account, Currency: ReservableCurrency<Account>>(
 pub fn free_deposit<Account, Currency: ReservableCurrency<Account>>(deposit: &Deposit<Account, Currency::Balance>) {
 	let err_amount = Currency::unreserve(&deposit.owner, deposit.amount);
 	debug_assert!(err_amount.is_zero());
-}
-
-// This code was copied from https://github.com/paritytech/substrate/blob/ded44948e2d5a398abcb4e342b0513cb690961bb/frame/transaction-payment/src/types.rs#L113
-// It is needed because u128 cannot be serialized by serde out of the box.
-#[cfg(feature = "std")]
-mod serde_balance {
-	use serde::{Deserialize, Deserializer, Serializer};
-
-	pub fn serialize<S: Serializer, T: std::fmt::Display>(t: &T, serializer: S) -> Result<S::Ok, S::Error> {
-		serializer.serialize_str(&t.to_string())
-	}
-
-	pub fn deserialize<'de, D: Deserializer<'de>, T: std::str::FromStr>(deserializer: D) -> Result<T, D::Error> {
-		let s = String::deserialize(deserializer)?;
-		s.parse::<T>()
-			.map_err(|_| serde::de::Error::custom("Parse from string failed"))
-	}
 }
 
 #[cfg(test)]
